@@ -1,20 +1,12 @@
-const React = require('react');
-const ReactDOM = require('react-dom');
-const Router = require('react-router');
-const History = require('history');
-const Routes = require('../core/Routes');
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { BrowserRouter } from 'react-router-dom';
+import BodyContent from '../core/BodyContent';
 
-const appHistory = Router.useRouterHistory(History.createHistory)();
 const container = document.createElement('div');
 document.body.appendChild(container);
 
 ReactDOM.render(
-  React.createElement(
-    Router.Router,
-    {
-      history: appHistory
-    },
-    Routes
-  ),
+  <BrowserRouter><BodyContent /></BrowserRouter>,
   container
 );
